+++ all-is-cubes/benches/raycast.rs
@@ -3,7 +3,7 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 use all_is_cubes::math::{Cube, Face7, GridAab};
-use all_is_cubes::raycast::{AxisAlignedRaycaster, Raycaster};
+use all_is_cubes::raycast::{AaRay, Raycaster};
 
 criterion_main!(benches);
 criterion_group! {
@@ -82,34 +82,35 @@ fn axis_aligned_raycaster_bench(c: &mut Criterion) {
     let mut group = c.benchmark_group("aa");
 
     group.bench_function("single step +X", |b| {
-        let mut raycaster =
-            AxisAlignedRaycaster::new(black_box(Cube::new(0, 0, 0)), black_box(Face7::PX));
+        let mut raycaster = black_box(AaRay::new(Cube::new(0, 0, 0), Face7::PX)).cast();
         b.iter(|| raycaster.next().unwrap());
     });
 
     group.bench_function("initialization", |b| {
-        b.iter(|| AxisAlignedRaycaster::new(black_box(Cube::new(0, 0, 0)), black_box(Face7::PX)))
+        b.iter(|| black_box(AaRay::new(Cube::new(0, 0, 0), Face7::PX)).cast())
     });
 
     group.bench_function("initialization with bounds (inside)", |b| {
         b.iter(|| {
-            AxisAlignedRaycaster::new(black_box(Cube::new(0, 0, 0)), black_box(Face7::PX))
+            black_box(AaRay::new(Cube::new(0, 0, 0), Face7::PX))
+                .cast()
                 .within(GridAab::from_lower_size([-1, -2, -3], [4, 5, 6]))
         })
     });
 
     group.bench_function("initialization with bounds (outside)", |b| {
         b.iter(|| {
-            AxisAlignedRaycaster::new(black_box(Cube::new(0, 0, 0)), black_box(Face7::PX))
+            black_box(AaRay::new(Cube::new(0, 0, 0), Face7::PX))
+                .cast()
                 .within(GridAab::from_lower_size([101, 102, 103], [4, 5, 6]))
         })
     });
 
     group.bench_function("many steps inside bounds", |b| {
         b.iter(|| {
-            let mut raycaster =
-                AxisAlignedRaycaster::new(black_box(Cube::new(0, 0, 0)), black_box(Face7::PX))
-                    .within(GridAab::from_lower_size([0, 0, 0], [100, 1000, 1000]));
+            let mut raycaster = black_box(AaRay::new(Cube::new(0, 0, 0), Face7::PX))
+                .cast()
+                .within(GridAab::from_lower_size([0, 0, 0], [100, 1000, 1000]));
             for _ in 1..100 {
                 black_box(raycaster.next());
             }
@@ -118,9 +119,9 @@ fn axis_aligned_raycaster_bench(c: &mut Criterion) {
 
     group.bench_function("many steps from outside bounds", |b| {
         b.iter(|| {
-            let mut raycaster =
-                AxisAlignedRaycaster::new(black_box(Cube::new(-50, 0, 0)), black_box(Face7::PX))
-                    .within(GridAab::from_lower_size([0, 0, 0], [100, 1000, 1000]));
+            let mut raycaster = black_box(AaRay::new(Cube::new(-50, 0, 0), Face7::PX))
+                .cast()
+                .within(GridAab::from_lower_size([0, 0, 0], [100, 1000, 1000]));
             for _ in 1..100 {
                 black_box(raycaster.next());
             }